User Rest Controller - added a users rest controller with post, put, delete, get by id, get by email. - the update method will not create an object if none is available. - other minor changes.

+++ src/main/java/leolem/demo/business/BorrowService.java
@@ -13,7 +13,7 @@ public class BorrowService {
   private final UserRepository userRepository;
   private final BookRepository bookRepository;
 
-  public void borrow(int bookId, int userId) throws EntityNotFoundException {
+  public void borrow(long bookId, long userId) throws EntityNotFoundException {
     User user = userRepository
         .findById(userId)
         .orElseThrow(() -> new EntityNotFoundException("Cannot find User with id: " + bookId));
